Add option to build the scanned synthesis opcodes
Because these opcodes are non-free according to debian they need to strip
scansyn* from the sources. They also need a patch to remove the opcodes,
that needs to be refreshed on each new csound release.

With this option they can drop the patch and simply specify that scansyn
opcodes are not to be built.