Add code to handle extra pfields (outside the PMAX limit) for the general case (insert_score_event_at_sample) and the corresponding API call (csoundScoreEventInternal).
Also supporting code for the 'passign' opcode to access the extra
memory block when called with an array as output.

There seem to be unnecessary allocations in the call chain which could
be optimized if 'insert_score_event_at_sample' could be notified that
the extra memory block (evt.c.extra) can be moved from the original
event instead of copied.