Handle quoted strings in option reader
When reading an option like `--env:FOO="bar baz"`, the first quotation
mark would be replaced by ETX while the second quotation mark would get
replaced by a NULL.  This has the effect of appending an ETX to the start
of the variable's value.

As a first pass to addressing the parsing of quoted strings, this adds
a post-processing step once we've detected the end of the current argument
which compacts down the string, removing all the special characters we
inserted while finding the end of the string.