Fix test_declare.csd test
The hangup with this test was in the declare statement.  This is fixed by
using a similar trick to udo definitions - flex start conditions.  When we
see the declare token, activate an exclusive start condition, and then
immediately relax it once we see the identifier token go by.

All tests now pass.